@@ -4,7 +4,7 @@
   "description": "Send an email with SendGrid when an update is made in Notion.",
   "main": "index.js",
   "engines": {
-    "node": "12.x"
+    "node": ">=14"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
@@ -12,7 +12,7 @@
   },
   "dependencies": {
     "@notionhq/client": "file:../../",
-    "@sendgrid/mail": "^7.7.0",
+    "@sendgrid/mail": "^8.1.5",
     "dotenv": "^16.0.1",
     "ts-node": "^10.8.2"
   },
