@@ -2,13 +2,13 @@
   "name": "notion-task-github-pr-sync",
   "version": "1.0.0",
   "description": "Sync Notion tasks with Github PRs",
-  "main": "index.js",
+  "main": "index.ts",
   "engines": {
-    "node": "12.x"
+    "node": ">=14"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": ""
+    "start": "ts-node index.ts"
   },
   "author": "Sid Verma",
   "license": "MIT",
@@ -17,5 +17,11 @@
     "dotenv": "^16.0.1",
     "lodash": "^4.17.21",
     "octokit": "^2.0.3"
+  },
+  "devDependencies": {
+    "@types/lodash": "^4.14.170",
+    "@types/node": "^18.0.0",
+    "ts-node": "^10.8.2",
+    "typescript": "^4.7.4"
   }
 }
