@@ -1,16 +1,17 @@
 /* ================================================================================
 
 	notion-github-sync.
-  
+
   Glitch example: https://glitch.com/edit/#!/notion-github-sync
   Find the official Notion API client @ https://github.com/makenotion/notion-sdk-js/
 
 ================================================================================ */
 
-const { Client } = require("@notionhq/client")
-const dotenv = require("dotenv")
-const { Octokit } = require("octokit")
-const _ = require("lodash")
+import { Client, isFullDatabase } from "@notionhq/client"
+import dotenv from "dotenv"
+import { Octokit } from "octokit"
+import _ from "lodash"
+import type { PaginatingEndpoints } from "@octokit/plugin-paginate-rest"
 
 dotenv.config()
 const octokit = new Octokit({ auth: process.env.GITHUB_KEY })
@@ -29,19 +30,19 @@ const gitHubIssuesIdToNotionPageId = {}
  * Initialize local data store.
  * Then sync with GitHub.
  */
-setInitialGitHubToNotionIdMap().then(syncNotionDatabaseWithGitHub)
+setInitialGitHubToNotionIdMap().then(syncNotionDataSourceWithGitHub)
 
 /**
  * Get and set the initial data store with issues currently in the database.
  */
 async function setInitialGitHubToNotionIdMap() {
-  const currentIssues = await getIssuesFromNotionDatabase()
+  const currentIssues = await getIssuesFromNotionDataSource()
   for (const { pageId, issueNumber } of currentIssues) {
     gitHubIssuesIdToNotionPageId[issueNumber] = pageId
   }
 }
 
-async function syncNotionDatabaseWithGitHub() {
+async function syncNotionDataSourceWithGitHub() {
   // Get all issues currently in the provided GitHub repository.
   console.log("\nFetching issues from GitHub repository...")
   const issues = await getGitHubIssuesForRepository()
@@ -67,12 +68,22 @@ async function syncNotionDatabaseWithGitHub() {
  *
  * @returns {Promise<Array<{ pageId: string, issueNumber: number }>>}
  */
-async function getIssuesFromNotionDatabase() {
+async function getIssuesFromNotionDataSource() {
   const pages = []
   let cursor = undefined
-  while (true) {
-    const { results, next_cursor } = await notion.databases.query({
-      database_id: databaseId,
+
+  const database = await notion.databases.retrieve({
+    database_id: String(databaseId),
+  })
+  if (!isFullDatabase(database)) {
+    console.error(`No read permissions on database: ${databaseId}`)
+    return []
+  }
+
+  const shouldContinue = true
+  while (shouldContinue) {
+    const { results, next_cursor } = await notion.dataSources.query({
+      database_id: database.data_sources[0].id,
       start_cursor: cursor,
     })
     pages.push(...results)
@@ -90,6 +101,10 @@ async function getIssuesFromNotionDatabase() {
       page_id: page.id,
       property_id: issueNumberPropertyId,
     })
+    if (!("number" in propertyResult)) {
+      console.error(`Issue number property not found for page: ${page.id}`)
+      continue
+    }
     issues.push({
       pageId: page.id,
       issueNumber: propertyResult.number,
@@ -109,21 +124,27 @@ async function getIssuesFromNotionDatabase() {
  */
 async function getGitHubIssuesForRepository() {
   const issues = []
-  const iterator = octokit.paginate.iterator(octokit.rest.issues.listForRepo, {
-    owner: process.env.GITHUB_REPO_OWNER,
-    repo: process.env.GITHUB_REPO_NAME,
-    state: "all",
-    per_page: 100,
-  })
+  const iterator = octokit.paginate.iterator(
+    octokit.rest.issues.listForRepo as unknown as keyof PaginatingEndpoints,
+    {
+      owner: process.env.GITHUB_REPO_OWNER,
+      repo: process.env.GITHUB_REPO_NAME,
+      state: "all",
+      per_page: 100,
+    }
+  )
   for await (const { data } of iterator) {
     for (const issue of data) {
-      if (!issue.pull_request) {
+      const castIssue: Record<string, unknown> =
+        typeof issue === "object" ? issue : {}
+
+      if (!castIssue.pull_request) {
         issues.push({
-          number: issue.number,
-          title: issue.title,
-          state: issue.state,
-          comment_count: issue.comments,
-          url: issue.html_url,
+          number: castIssue.number,
+          title: castIssue.title,
+          state: castIssue.state,
+          comment_count: castIssue.comments,
+          url: castIssue.html_url,
         })
       }
     }
@@ -190,9 +211,9 @@ async function updatePages(pagesToUpdate) {
   const pagesToUpdateChunks = _.chunk(pagesToUpdate, OPERATION_BATCH_SIZE)
   for (const pagesToUpdateBatch of pagesToUpdateChunks) {
     await Promise.all(
-      pagesToUpdateBatch.map(({ pageId, ...issue }) =>
+      pagesToUpdateBatch.map(issue =>
         notion.pages.update({
-          page_id: pageId,
+          page_id: Object(issue).pageId,
           properties: getPropertiesFromIssue(issue),
         })
       )
@@ -214,7 +235,7 @@ function getPropertiesFromIssue(issue) {
   const { title, number, state, comment_count, url } = issue
   return {
     Name: {
-      title: [{ type: "text", text: { content: title } }],
+      title: [{ type: "text" as const, text: { content: title } }],
     },
     "Issue Number": {
       number,
