@@ -8,28 +8,45 @@ const apiKey = process.env.NOTION_API_KEY
 
 const notion = new Client({ auth: apiKey })
 
-/* 
+/*
 ---------------------------------------------------------------------------
 */
 
 /**
  * Resources:
  * - Appending block children endpoint (notion.blocks.children.append(): https://developers.notion.com/reference/patch-block-children)
+ * - Rich text options: https://developers.notion.com/reference/rich-text
  * - Working with page content guide: https://developers.notion.com/docs/working-with-page-content
  */
 
 async function main() {
   const blockId = pageId // Blocks can be appended to other blocks *or* pages. Therefore, a page ID can be used for the block_id parameter
-  const newHeadingResponse = await notion.blocks.children.append({
+  const linkedTextResponse = await notion.blocks.children.append({
     block_id: blockId,
     // Pass an array of blocks to append to the page: https://developers.notion.com/reference/block#block-type-objects
     children: [
       {
-        heading_2: {
+        heading_3: {
           rich_text: [
             {
               text: {
-                content: "Types of kale", // This is the text that will be displayed in Notion
+                content: "Tuscan  kale", // This is the text that will be displayed in Notion
+              },
+            },
+          ],
+        },
+      },
+      {
+        paragraph: {
+          rich_text: [
+            {
+              text: {
+                content:
+                  "Tuscan  kale is a variety of kale with a long tradition in Italian cuisine, especially that of Tuscany. It is also known as Tuscan kale, Italian kale, dinosaur kale, kale, flat back kale, palm tree kale, or black Tuscan palm.",
+                link: {
+                  // Include a url to make the paragraph a link in Notion
+                  url: "https://en.wikipedia.org/wiki/Kale",
+                },
               },
             },
           ],
@@ -39,7 +56,7 @@ async function main() {
   })
 
   // Print the new block(s) response
-  console.log(newHeadingResponse)
+  console.log(JSON.stringify(linkedTextResponse, null, 2))
 }
 
 main()
