@@ -8,7 +8,7 @@ const apiKey = process.env.NOTION_API_KEY
 
 const notion = new Client({ auth: apiKey })
 
-/* 
+/*
 ---------------------------------------------------------------------------
 */
 
@@ -64,7 +64,7 @@ async function main() {
   })
 
   // Print the new block(s) response
-  console.log(styledLinkTextResponse)
+  console.log(JSON.stringify(styledLinkTextResponse, null, 2))
 }
 
 main()
