@@ -1,4 +1,4 @@
-import { Client } from "@notionhq/client"
+import { Client, isFullDatabase } from "@notionhq/client"
 import { config } from "dotenv"
 import { propertiesForNewPages } from "./sampleData.js"
 
@@ -9,7 +9,7 @@ const apiKey = process.env.NOTION_API_KEY
 
 const notion = new Client({ auth: apiKey })
 
-/* 
+/*
 ---------------------------------------------------------------------------
 */
 
@@ -20,42 +20,35 @@ const notion = new Client({ auth: apiKey })
  * - Working with databases guide: https://developers.notion.com/docs/working-with-databases
  * Query a database: https://developers.notion.com/reference/post-database-query
  * Filter database entries: https://developers.notion.com/reference/post-database-query-filter
- * Sort database entries: https://developers.notion.com/reference/post-database-query-sort
  */
 
-async function addNotionPageToDatabase(databaseId, pageProperties) {
+async function addNotionPageToDataSource(dataSourceId, pageProperties) {
   await notion.pages.create({
     parent: {
-      database_id: databaseId,
+      data_source_id: dataSourceId,
     },
     properties: pageProperties, // Note: Page properties must match the schema of the database
   })
 }
 
-async function queryAndSortDatabase(databaseId) {
+async function queryDataSource(dataSourceId) {
   console.log("Querying database...")
-  // This query will filter and sort database entries. The returned pages will have a "Last ordered" property that is more recent than 2022-12-31. Any database property can be filtered or sorted. Pass multiple sort objects to the "sorts" array to apply more than one sorting rule.
-  const lastOrderedIn2023Alphabetical = await notion.databases.query({
-    database_id: databaseId,
+  // This query will filter database entries and return pages that have a "Last ordered"
+  // property that is more recent than 2022-12-31. Use multiple filters with the AND/OR
+  // options: https://developers.notion.com/reference/post-database-query-filter.
+  const lastOrderedIn2023 = await notion.dataSources.query({
+    database_id: dataSourceId,
     filter: {
       property: "Last ordered",
       date: {
         after: "2022-12-31",
       },
     },
-    sorts: [
-      {
-        property: "Grocery item",
-        direction: "descending",
-      },
-    ],
   })
 
-  // Print filtered/sorted results
-  console.log(
-    'Pages with the "Last ordered" date after 2022-12-31 in descending order:'
-  )
-  console.log(lastOrderedIn2023Alphabetical)
+  // Print filtered results
+  console.log('Pages with the "Last ordered" date after 2022-12-31:')
+  console.log(JSON.stringify(lastOrderedIn2023, null, 2))
 }
 
 async function main() {
@@ -73,38 +66,47 @@ async function main() {
         },
       },
     ],
-    properties: {
-      // These properties represent columns in the database (i.e. its schema)
-      "Grocery item": {
-        type: "title",
-        title: {},
-      },
-      Price: {
-        type: "number",
-        number: {
-          format: "dollar",
+    initial_data_source: {
+      properties: {
+        // These properties represent columns in the data source (i.e. its schema)
+        "Grocery item": {
+          type: "title",
+          title: {},
         },
-      },
-      "Last ordered": {
-        type: "date",
-        date: {},
+        Price: {
+          type: "number",
+          number: {
+            format: "dollar",
+          },
+        },
+        "Last ordered": {
+          type: "date",
+          date: {},
+        },
       },
     },
   })
-  // Print the new database's URL. Visit the URL in your browser to see the pages that get created in the next step.
-  console.log(newDatabase.url)
 
-  const databaseId = newDatabase.id
-  if (!databaseId) return
+  if (!isFullDatabase(newDatabase)) {
+    console.error(`No read permissions on database: ${newDatabase.id}`)
+    return
+  }
+
+  // Print the new database's URL. Visit the URL in your browser to see the pages that get created in the next step.
+  console.log(newDatabase.id)
+
+  const dataSourceId = newDatabase.data_sources[0].id
+  // If there is no ID (if there's an error), return.
+  if (!dataSourceId) return
 
   console.log("Adding new pages...")
   for (let i = 0; i < propertiesForNewPages.length; i++) {
     // Add a few new pages to the database that was just created
-    await addNotionPageToDatabase(databaseId, propertiesForNewPages[i])
+    await addNotionPageToDataSource(dataSourceId, propertiesForNewPages[i])
   }
 
-  // After adding pages, query the database entries (pages) and sort the results
-  queryAndSortDatabase(databaseId)
+  // After adding pages, query the database entries (pages)
+  queryDataSource(dataSourceId)
 }
 
 main()
