@@ -28,21 +28,18 @@ import {
   type ListBlockChildrenParameters,
   type ListBlockChildrenResponse,
   listBlockChildren,
-  type ListDatabasesParameters,
-  type ListDatabasesResponse,
-  listDatabases,
-  type GetDatabaseParameters,
-  type GetDatabaseResponse,
-  getDatabase,
-  type QueryDatabaseParameters,
-  type QueryDatabaseResponse,
-  queryDatabase,
-  type CreateDatabaseParameters,
-  type CreateDatabaseResponse,
-  createDatabase,
-  type UpdateDatabaseParameters,
-  type UpdateDatabaseResponse,
-  updateDatabase,
+  type QueryDataSourceParameters,
+  type QueryDataSourceResponse,
+  queryDataSource,
+  type CreateDataSourceParameters,
+  type CreateDataSourceResponse,
+  createDataSource,
+  type UpdateDataSourceParameters,
+  type UpdateDataSourceResponse,
+  updateDataSource,
+  type GetDataSourceParameters,
+  type GetDataSourceResponse,
+  getDataSource,
   type CreatePageParameters,
   type CreatePageResponse,
   createPage,
@@ -100,6 +97,15 @@ import {
   type ListFileUploadsParameters,
   type ListFileUploadsResponse,
   listFileUploads,
+  GetDatabaseParameters,
+  GetDatabaseResponse,
+  getDatabase,
+  CreateDatabaseResponse,
+  CreateDatabaseParameters,
+  createDatabase,
+  UpdateDatabaseParameters,
+  UpdateDatabaseResponse,
+  updateDatabase,
 } from "./api-endpoints"
 import {
   version as PACKAGE_VERSION,
@@ -374,23 +380,6 @@ export default class Client {
 
   public readonly databases = {
     /**
-     * List databases
-     *
-     * @deprecated Please use `search`
-     */
-    list: (
-      args: WithAuth<ListDatabasesParameters>
-    ): Promise<ListDatabasesResponse> => {
-      return this.request<ListDatabasesResponse>({
-        path: listDatabases.path(),
-        method: listDatabases.method,
-        query: pick(args, listDatabases.queryParams),
-        body: pick(args, listDatabases.bodyParams),
-        auth: args?.auth,
-      })
-    },
-
-    /**
      * Retrieve a database
      */
     retrieve: (
@@ -401,21 +390,6 @@ export default class Client {
         method: getDatabase.method,
         query: pick(args, getDatabase.queryParams),
         body: pick(args, getDatabase.bodyParams),
-        auth: args?.auth,
-      })
-    },
-
-    /**
-     * Query a database
-     */
-    query: (
-      args: WithAuth<QueryDatabaseParameters>
-    ): Promise<QueryDatabaseResponse> => {
-      return this.request<QueryDatabaseResponse>({
-        path: queryDatabase.path(args),
-        method: queryDatabase.method,
-        query: pick(args, queryDatabase.queryParams),
-        body: pick(args, queryDatabase.bodyParams),
         auth: args?.auth,
       })
     },
@@ -446,6 +420,68 @@ export default class Client {
         method: updateDatabase.method,
         query: pick(args, updateDatabase.queryParams),
         body: pick(args, updateDatabase.bodyParams),
+        auth: args?.auth,
+      })
+    },
+  }
+
+  public readonly dataSources = {
+    /**
+     * Retrieve a data source
+     */
+    retrieve: (
+      args: WithAuth<GetDataSourceParameters>
+    ): Promise<GetDataSourceResponse> => {
+      return this.request<GetDataSourceResponse>({
+        path: getDataSource.path(args),
+        method: getDataSource.method,
+        query: pick(args, getDataSource.queryParams),
+        body: pick(args, getDataSource.bodyParams),
+        auth: args?.auth,
+      })
+    },
+
+    /**
+     * Query a data source
+     */
+    query: (
+      args: WithAuth<QueryDataSourceParameters>
+    ): Promise<QueryDataSourceResponse> => {
+      return this.request<QueryDataSourceResponse>({
+        path: queryDataSource.path(args),
+        method: queryDataSource.method,
+        query: pick(args, queryDataSource.queryParams),
+        body: pick(args, queryDataSource.bodyParams),
+        auth: args?.auth,
+      })
+    },
+
+    /**
+     * Create a data source
+     */
+    create: (
+      args: WithAuth<CreateDataSourceParameters>
+    ): Promise<CreateDataSourceResponse> => {
+      return this.request<CreateDataSourceResponse>({
+        path: createDataSource.path(),
+        method: createDataSource.method,
+        query: pick(args, createDataSource.queryParams),
+        body: pick(args, createDataSource.bodyParams),
+        auth: args?.auth,
+      })
+    },
+
+    /**
+     * Update a data source
+     */
+    update: (
+      args: WithAuth<UpdateDataSourceParameters>
+    ): Promise<UpdateDataSourceResponse> => {
+      return this.request<UpdateDataSourceResponse>({
+        path: updateDataSource.path(args),
+        method: updateDataSource.method,
+        query: pick(args, updateDataSource.queryParams),
+        body: pick(args, updateDataSource.bodyParams),
         auth: args?.auth,
       })
     },
