@@ -10,28 +10,58 @@
 export type {
   AppendBlockChildrenParameters,
   AppendBlockChildrenResponse,
+  AudioBlockObjectResponse,
   BlockObjectRequest,
   BlockObjectResponse,
+  BookmarkBlockObjectResponse,
+  BotUserObjectResponse,
+  BreadcrumbBlockObjectResponse,
+  BulletedListItemBlockObjectResponse,
+  ButtonPropertyItemObjectResponse,
+  CalloutBlockObjectResponse,
+  CheckboxPropertyItemObjectResponse,
+  ChildDatabaseBlockObjectResponse,
+  ChildPageBlockObjectResponse,
+  CodeBlockObjectResponse,
+  ColumnBlockObjectResponse,
+  ColumnListBlockObjectResponse,
   CommentObjectResponse,
   CompleteFileUploadParameters,
   CompleteFileUploadResponse,
   CreateCommentParameters,
   CreateCommentResponse,
-  GetCommentParameters,
-  GetCommentResponse,
   CreateDatabaseParameters,
   CreateDatabaseResponse,
+  CreateDataSourceParameters,
+  CreateDataSourceResponse,
+  CreatedByPropertyItemObjectResponse,
+  CreatedTimePropertyItemObjectResponse,
   CreateFileUploadParameters,
   CreateFileUploadResponse,
   CreatePageParameters,
   CreatePageResponse,
   DatabaseObjectResponse,
+  DataSourceObjectResponse,
+  DatePropertyItemObjectResponse,
+  DeleteBlockParameters,
+  DeleteBlockResponse,
+  DividerBlockObjectResponse,
+  EmailPropertyItemObjectResponse,
+  EmbedBlockObjectResponse,
+  EquationBlockObjectResponse,
   EquationRichTextItemResponse,
+  FileBlockObjectResponse,
+  FilesPropertyItemObjectResponse,
   FileUploadObjectResponse,
+  FormulaPropertyItemObjectResponse,
   GetBlockParameters,
   GetBlockResponse,
+  GetCommentParameters,
+  GetCommentResponse,
   GetDatabaseParameters,
   GetDatabaseResponse,
+  GetDataSourceParameters,
+  GetDataSourceResponse,
   GetFileUploadParameters,
   GetFileUploadResponse,
   GetPageParameters,
@@ -42,13 +72,27 @@ export type {
   GetSelfResponse,
   GetUserParameters,
   GetUserResponse,
+  GroupObjectResponse,
+  Heading1BlockObjectResponse,
+  Heading2BlockObjectResponse,
+  Heading3BlockObjectResponse,
+  ImageBlockObjectResponse,
+  LastEditedByPropertyItemObjectResponse,
+  LastEditedTimePropertyItemObjectResponse,
+  LinkPreviewBlockObjectResponse,
+  LinkToPageBlockObjectResponse,
   ListBlockChildrenParameters,
   ListBlockChildrenResponse,
   ListCommentsParameters,
   ListCommentsResponse,
   ListFileUploadsParameters,
   ListFileUploadsResponse,
+  ListUsersParameters,
+  ListUsersResponse,
   MentionRichTextItemResponse,
+  MultiSelectPropertyItemObjectResponse,
+  NumberedListItemBlockObjectResponse,
+  NumberPropertyItemObjectResponse,
   OauthIntrospectParameters,
   OauthIntrospectResponse,
   OauthRevokeParameters,
@@ -58,20 +102,53 @@ export type {
   PageObjectResponse,
   ParagraphBlockObjectResponse,
   PartialBlockObjectResponse,
+  PartialCommentObjectResponse,
   PartialDatabaseObjectResponse,
+  PartialDataSourceObjectResponse,
   PartialPageObjectResponse,
   PartialUserObjectResponse,
+  PdfBlockObjectResponse,
+  PeoplePropertyItemObjectResponse,
+  PersonUserObjectResponse,
+  PhoneNumberPropertyItemObjectResponse,
+  PropertyItemListResponse,
+  PropertyItemObjectResponse,
+  QueryDataSourceParameters,
+  QueryDataSourceResponse,
+  QuoteBlockObjectResponse,
+  RelationPropertyItemObjectResponse,
   RichTextItemResponse,
+  RichTextPropertyItemObjectResponse,
+  RollupPropertyItemObjectResponse,
+  SearchParameters,
+  SearchResponse,
+  SelectPropertyItemObjectResponse,
   SendFileUploadParameters,
   SendFileUploadResponse,
+  StatusPropertyItemObjectResponse,
+  SyncedBlockBlockObjectResponse,
+  TableBlockObjectResponse,
+  TableOfContentsBlockObjectResponse,
+  TableRowBlockObjectResponse,
+  TemplateBlockObjectResponse,
   TextRichTextItemResponse,
+  TitlePropertyItemObjectResponse,
+  ToDoBlockObjectResponse,
+  ToggleBlockObjectResponse,
+  UniqueIdPropertyItemObjectResponse,
+  UnsupportedBlockObjectResponse,
   UpdateBlockParameters,
   UpdateBlockResponse,
   UpdateDatabaseParameters,
   UpdateDatabaseResponse,
+  UpdateDataSourceParameters,
+  UpdateDataSourceResponse,
   UpdatePageParameters,
   UpdatePageResponse,
+  UrlPropertyItemObjectResponse,
   UserObjectResponse,
+  VerificationPropertyItemObjectResponse,
+  VideoBlockObjectResponse,
 } from "./api-endpoints"
 export { default as Client } from "./Client"
 export { LogLevel, Logger } from "./logging"
@@ -92,9 +169,10 @@ export {
   collectPaginatedAPI,
   iteratePaginatedAPI,
   isFullBlock,
+  isFullDataSource,
   isFullDatabase,
   isFullPage,
   isFullUser,
   isFullComment,
-  isFullPageOrDatabase,
+  isFullPageOrDataSource,
 } from "./helpers"
